Move activity recognition repo interface and models to domain package
