+++ app/src/main/java/com/stochanskyi/librariesdemo/app/di/modules/RepositoryBindsModule.kt
@@ -1,8 +1,8 @@
 package com.stochanskyi.librariesdemo.app.di.modules
 
-import com.stochanskyi.librariesdemo.data.repository.activityrecognition.ActivityRecognitionRepository
 import com.stochanskyi.librariesdemo.data.repository.activityrecognition.ActivityRecognitionRepositoryImpl
 import com.stochanskyi.librariesdemo.data.repository.simplecall.SimpleCallRepositoryImpl
+import com.stochanskyi.librariesdemo.domain.features.activityrecognition.ActivityRecognitionRepository
 import com.stochanskyi.librariesdemo.domain.features.simplecall.SimpleCallRepository
 import dagger.Binds
 import dagger.Module
