--- app/src/main/java/com/stochanskyi/librariesdemo/data/repository/activityrecognition/models/UserActivity.kt
@@ -1,8 +1,0 @@
-package com.stochanskyi.librariesdemo.data.repository.activityrecognition.models
-
-import com.stochanskyi.librariesdemo.data.repository.activityrecognition.models.type.ActivityType
-
-class UserActivity(
-    val activityType: ActivityType,
-    val confidence: Int
-)
