+++ app/src/main/java/com/stochanskyi/librariesdemo/domain/features/activityrecognition/models/UserActivity.kt
@@ -1,0 +1,8 @@
+package com.stochanskyi.librariesdemo.domain.features.activityrecognition.models
+
+import com.stochanskyi.librariesdemo.domain.features.activityrecognition.models.type.ActivityType
+
+class UserActivity(
+    val activityType: ActivityType,
+    val confidence: Int
+)
