+++ app/src/main/java/com/stochanskyi/librariesdemo/domain/features/activityrecognition/models/type/ActivityType.kt
@@ -1,4 +1,4 @@
-package com.stochanskyi.librariesdemo.data.repository.activityrecognition.models.type
+package com.stochanskyi.librariesdemo.domain.features.activityrecognition.models.type
 
 interface ActivityType {
     fun <T> transform(transformer: ActivityTypeTransformer<T>): T
